Program Crashing Bug Fix
there was an issue with how the program writes the information tow the
daysimeter, it would not include a 0 with the month or the day if the
number was less then 10.
If a daysimeter was started before this patch, what you need to do is
edit the file "log_info.txt" to include a '0' before the month if the
month is less then 10.