Updates and Name Change
This update will allow the user to know how long the Daysimeter will be
able to run and collect data. This is also explained in the README file.
To make the name smaller the Setup file was renamed to DSS setup, this
allows the program to be created much easier. When installed the program
will still be called Daysimeter Start_Stop.